+++ src/Protocol/CMakeLists.txt
@@ -34,5 +34,4 @@ target_sources(
 )
 
 add_subdirectory(Palettes)
-add_subdirectory(Enums)
 add_subdirectory(Sounds)
