--- src/Protocol/Enums/CMakeLists.txt
@@ -1,5 +1,0 @@
-target_sources(
-	${CMAKE_PROJECT_NAME} PRIVATE
-
-	SoundEvent.h
-)
