+++ src/Protocol/Sounds/CMakeLists.txt
@@ -1,5 +1,5 @@
 target_sources(
 	${CMAKE_PROJECT_NAME} PRIVATE
 
-	Protocol_1_8.cpp
+	PSounds_1_8.cpp
 )
